requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
python-decouple==3.3
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
